Fix Average/N value code and answer in pdf

+++ Year2/Data_Structures_6006/Lab3/ArrayTester.java
@@ -18,7 +18,7 @@ public class ArrayTester {
             }
 
             long avgTime = totalTime / 5;
-            System.out.println(" \t" + avgTime + " \t" + avgTime / N);
+            System.out.println(" \t" + avgTime + " \t" + (float) avgTime / N);
         }
     }
 }
